refactor(navigationfooter): rename to NavigationFooter
