+++ packages/components/src/index.ts
@@ -45,8 +45,6 @@ export * from "./core/Dropdown";
 export { default as Dropdown } from "./core/Dropdown";
 export * from "./core/DropdownMenu";
 export { default as DropdownMenu } from "./core/DropdownMenu";
-export * from "./core/Footer";
-export { default as Footer } from "./core/Footer";
 export * from "./core/Icon";
 export { default as Icon } from "./core/Icon";
 export * from "./core/InputCheckbox";
@@ -79,6 +77,8 @@ export * from "./core/MenuItem";
 export { default as MenuItem } from "./core/MenuItem";
 export * from "./core/MenuSelect";
 export { default as MenuSelect } from "./core/MenuSelect";
+export * from "./core/NavigationFooter";
+export { default as Footer } from "./core/NavigationFooter";
 export * from "./core/NavigationJumpTo";
 export { default as NavigationJumpTo } from "./core/NavigationJumpTo";
 export * from "./core/Notification";
